assesment enhancement

+++ system_design/FileManagement/src/main/java/com/demo/FileManagement/model/FileDB.java
@@ -1,20 +1,20 @@
 package com.demo.FileManagement.model;
 
 
-import javax.persistence.CascadeType;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.Lob;
-import javax.persistence.OneToOne;
-import javax.persistence.Table;
 
 import org.hibernate.annotations.GenericGenerator;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.Lob;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
 import lombok.Data;
 
 @Entity
@@ -28,7 +28,7 @@ public class FileDB {
   
   private String fileName;
   
-  private Long fileSize;
+  private Float fileSize;
 
   @JsonIgnore
   @OneToOne(fetch = FetchType.LAZY,cascade = CascadeType.ALL)
